+++ src/components/BlogPostDisplay.tsx
@@ -1,20 +1,26 @@
+import { useEntries } from "@/contexts/usePosts";
 import { useEffect, useState } from "react";
+import { type PostForIndex } from "types";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
-import { type PostWithEmbedding } from "types";
 
 type BlogPostDisplayProps = {
-  data: PostWithEmbedding;
+  data: PostForIndex;
 };
 
 export function BlogPostDisplay({ data }: BlogPostDisplayProps) {
   const [content, setContent] = useState<string>("");
   const [error, setError] = useState<string | null>(null);
+  const { selectedPost } = useEntries();
 
   useEffect(() => {
     let isMounted = true;
 
-    fetch(`/posts/${data.filename}`)
+    if (!selectedPost) {
+      return;
+    }
+
+    fetch(`/posts/${selectedPost.filename}`)
       .then((res) => res.text())
       .then((raw) => {
         const content = raw.replace(/^---[\s\S]*?---/, "").trim();
@@ -27,7 +33,7 @@ export function BlogPostDisplay({ data }: BlogPostDisplayProps) {
     return () => {
       isMounted = false;
     };
-  }, [data.filename]);
+  }, [selectedPost]);
 
   if (error) {
     return (
@@ -38,25 +44,25 @@ export function BlogPostDisplay({ data }: BlogPostDisplayProps) {
   }
 
   return (
-    <article className="w-full max-w-2xl mx-auto bg-white rounded-lg shadow border p-6 text-left">
-      <h1 className="text-xl md:text-2xl font-extrabold text-black mb-2 text-left">
+    <article className="w-full max-w-2xl mx-auto rounded-lg shadow border p-6 text-left">
+      <h1 className="text-xl md:text-2xl font-extrabold mb-2 text-left">
         {data.title}
       </h1>
       <div className="text-xs md:text-sm text-neutral-400 mb-6 text-left">
         {new Date(data.date).toLocaleDateString()}
       </div>
-      <div className="prose prose-neutral dark:prose-invert max-w-none text-left">
+      <div className="prose max-w-none text-left dark:prose-invert">
         <ReactMarkdown
           remarkPlugins={[remarkGfm]}
           components={{
             h1: ({ node, ...props }) => (
-              <h1 className="text-xl font-bold mt-6 mb-2" {...props} />
+              <h1 className="text-2xl font-semibold mt-6 mb-2" {...props} />
             ),
             h2: ({ node, ...props }) => (
-              <h2 className="text-lg font-bold mt-4 mb-2" {...props} />
+              <h2 className="text-xl font-semibold mt-4 mb-2" {...props} />
             ),
             h3: ({ node, ...props }) => (
-              <h3 className="text-base font-semibold mt-3 mb-1" {...props} />
+              <h3 className="text-lg font-semibold mt-3 mb-1" {...props} />
             ),
             h4: ({ node, ...props }) => (
               <h4 className="text-base font-medium mt-2 mb-1" {...props} />
@@ -74,12 +80,7 @@ export function BlogPostDisplay({ data }: BlogPostDisplayProps) {
               <ol className="list-decimal ml-6 my-2" {...props} />
             ),
             li: ({ node, ...props }) => <li className="my-1" {...props} />,
-            hr: ({ node, ...props }) => (
-              <hr
-                className="my-8 border-t border-neutral-300 dark:border-neutral-700"
-                {...props}
-              />
-            ),
+            hr: ({ node, ...props }) => <hr className="my-8 " {...props} />,
           }}
         >
           {content}